requirements.txt

# Core scientific stack
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
scipy>=1.11
scikit-image>=0.22
tqdm>=4.66
PyYAML>=6.0.1
python-dotenv>=1.0.1

# CV / DL
opencv-python>=4.9
ultralytics>=8.3.0,<9

# Video composition
moviepy>=1.0.3
Pillow>=10.0
